backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
pymysql==1.1.0
google-generativeai==0.4.0
pymupdf==1.23.22
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.2.0
cryptography==42.0.5
gunicorn==21.2.0
